+++ src/main/java/SouthPanel/SouthPanel.java
@@ -1,0 +1,12 @@
+package SouthPanel;
+
+import java.awt.Color;
+
+import javax.swing.JPanel;
+
+public class SouthPanel extends JPanel{	
+	public SouthPanel(){
+		setBackground(Color.WHITE);
+		add(new InputTextPanel());
+	}
+}
