+++ src/test/java/GameInterface/WordLabelArrayTest.java
@@ -5,43 +5,43 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Before;
 import org.junit.Test;
 
-import GameInterface.WordLabelArray;
-import GameInterface.WordLabel;
+import CenterPanel.FallWordLabel;
+import CenterPanel.FallWordLabelArray;
 
 public class WordLabelArrayTest {
 	
 	@Before
 	public void removeTest(){
-		WordLabel wla=new WordLabel("korean");
-		WordLabelArray.add(wla);
-		WordLabelArray.remove(0);
-		assertEquals(WordLabelArray.getNumOfLabel(),0);
+		FallWordLabel wla=new FallWordLabel("korean");
+		FallWordLabelArray.add(wla);
+		FallWordLabelArray.remove(0);
+		assertEquals(FallWordLabelArray.getNumOfLabel(),0);
 	}
 
 	@Test
 	public void getTest() {
-		WordLabel wla=new WordLabel("korean");
-		WordLabelArray.add(wla);
-		assertEquals(WordLabelArray.get(0),wla);
-		WordLabelArray.remove(0);
+		FallWordLabel wla=new FallWordLabel("korean");
+		FallWordLabelArray.add(wla);
+		assertEquals(FallWordLabelArray.get(0),wla);
+		FallWordLabelArray.remove(0);
 	}
 	
 	
 	@Test
 	public void getTextTest() {
-		WordLabel wla=new WordLabel("korean");
-		WordLabelArray.add(wla);
-		assertEquals(WordLabelArray.getText(0),wla.getText());
-		WordLabelArray.remove(0);
+		FallWordLabel wla=new FallWordLabel("korean");
+		FallWordLabelArray.add(wla);
+		assertEquals(FallWordLabelArray.getText(0),wla.getText());
+		FallWordLabelArray.remove(0);
 	}
 	
 	@Test
 	public void setTextTest() {
-		WordLabel wla=new WordLabel("korean");
-		WordLabelArray.add(wla);
-		WordLabelArray.setText(0,"english");
-		assertEquals(WordLabelArray.getText(0),"english");
-		WordLabelArray.remove(0);
+		FallWordLabel wla=new FallWordLabel("korean");
+		FallWordLabelArray.add(wla);
+		FallWordLabelArray.setText(0,"english");
+		assertEquals(FallWordLabelArray.getText(0),"english");
+		FallWordLabelArray.remove(0);
 	}
