+++ src/test/java/SouthPanel/SouthPanelTest.java
@@ -1,10 +1,10 @@
-package MainFrame;
+package SouthPanel;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-import MainFrame.SouthPanel;
+import SouthPanel.SouthPanel;
 
 public class SouthPanelTest {
